+++ src/Model/Term/TermRepository.php
@@ -76,8 +76,10 @@ abstract class TermRepository extends Repository
      */
     public function find($args = null)
     {
-        if (is_null($args)) {
-            $args = $this->newParams();
+        if (is_null($args) or is_array($args)) {
+            $defaults = $this->newParams();
+            $defaults = ($defaults instanceof Arrayable) ? $defaults->toArray() : $defaults;
+            $args     = is_array($args) ? array_merge_recursive($defaults, $args) : $defaults;
         }
 
         $args = ($args instanceof Arrayable) ? $args->toArray() : $args;