+++ src/Model/User/UserRepository.php
@@ -45,8 +45,10 @@ class UserRepository extends Repository implements UserRepositoryContract
 
     public function find($args = null)
     {
-        if (is_null($args)) {
-            $args = $this->newParams();
+        if (is_null($args) or is_array($args)) {
+            $defaults = $this->newParams();
+            $defaults = ($defaults instanceof Arrayable) ? $defaults->toArray() : $defaults;
+            $args     = is_array($args) ? array_merge_recursive($defaults, $args) : $defaults;
         }
 
         $args = ($args instanceof Arrayable) ? $args->toArray() : $args;